requirements.txt
aiohttp>=3.8.0
aiofiles>=0.8.0
schedule>=1.2.0
croniter>=1.4.1
torch>=2.0.0
torchvision>=0.15.0
wdtagger>=0.3.0
Pillow>=9.0.0
numpy>=1.21.0
timm>=0.9.12
transformers>=4.20.0
tqdm>=4.64.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
